@@ -1,15 +1,14 @@
 name: extended_image
 description: Official extension image, support placeholder(loading)/ failed state, cache network, zoom/pan, photo view, slide out page, editor(crop,rotate,flip), painting etc.
-version: 6.0.1
+version: 6.2.1
 homepage: https://github.com/fluttercandies/extended_image
 
 environment:
-  sdk: '>=2.14.0 <3.0.0'
-  flutter: '>=2.8.0'
+  sdk: '>=2.17.0 <3.0.0'
+  flutter: '>=3.0.0'
 
 dependencies:
-  extended_image_library: ^3.1.0
+  extended_image_library: ^3.3.0
   flutter:
     sdk: flutter
-  meta: ^1.1.8
-
+  meta: ^1.7.0
