@@ -1,7 +1,7 @@
 import 'package:extended_image/extended_image.dart';
 import 'package:flutter/material.dart';
 
-class FlatButtonWithIcon extends TextButton with MaterialButtonWithIconMixin {
+class FlatButtonWithIcon extends TextButton {
   FlatButtonWithIcon({
     Key? key,
     required VoidCallback onPressed,
