@@ -141,12 +141,13 @@ class _PhotoViewDemoState extends State<PhotoViewDemo> {
                                   content,
                                   onSpecialTextTap: (dynamic parameter) {
                                     if (parameter.toString().startsWith('\$')) {
-                                      launch(
-                                          'https://github.com/fluttercandies');
+                                      launchUrl(Uri.parse(
+                                          'https://github.com/fluttercandies'));
                                     } else if (parameter
                                         .toString()
                                         .startsWith('@')) {
-                                      launch('mailto:zmtzawqlp@live.com');
+                                      launchUrl(Uri.parse(
+                                          'mailto:zmtzawqlp@live.com'));
                                     }
                                   },
                                   specialTextSpanBuilder:
@@ -163,8 +164,8 @@ class _PhotoViewDemoState extends State<PhotoViewDemo> {
                                         InkWell(
                                           child: const Text('more'),
                                           onTap: () {
-                                            launch(
-                                                'https://github.com/fluttercandies/extended_text');
+                                            launchUrl(Uri.parse(
+                                                'https://github.com/fluttercandies/extended_text'));
                                           },
                                         )
                                       ],
