@@ -119,7 +119,8 @@ class MainPage extends StatelessWidget {
                 ),
               ),
               onPressed: () {
-                launch('https://github.com/fluttercandies/extended_image');
+                launchUrl(Uri.parse(
+                    'https://github.com/fluttercandies/extended_image'));
               },
             ),
           ),
@@ -131,7 +132,7 @@ class MainPage extends StatelessWidget {
                 child: Image.network(
                     'https://pub.idqqimg.com/wpa/images/group.png'),
                 onPressed: () {
-                  launch('https://jq.qq.com/?_wv=1027&k=5bcc0gy');
+                  launchUrl(Uri.parse('https://jq.qq.com/?_wv=1027&k=5bcc0gy'));
                 },
               ),
             )
