@@ -1,6 +1,7 @@
 library extended_image;
 
-export 'package:extended_image_library/extended_image_library.dart';
+export 'package:extended_image_library/extended_image_library.dart'
+    hide File, FileImage;
 
 export 'src/border_painter.dart';
 export 'src/editor/editor.dart';
