@@ -30,13 +30,14 @@ class ExtendedRenderImage extends RenderBox {
     bool matchTextDirection = false,
     TextDirection? textDirection,
     bool invertColors = false,
+    bool isAntiAlias = false,
     FilterQuality filterQuality = FilterQuality.low,
     Rect? sourceRect,
     AfterPaintImage? afterPaintImage,
     BeforePaintImage? beforePaintImage,
     GestureDetails? gestureDetails,
     EditActionDetails? editActionDetails,
-    bool isAntiAlias = false,
+  
   })  : _image = image,
         _width = width,
         _height = height,
